reorganized answers jsx
