+++ curricula/static/curricula/js/components/answers/multiple.jsx
