+++ curricula/static/curricula/js/components/answers/multiselect.jsx
