+++ curricula/static/curricula/style.css
@@ -518,6 +518,4 @@ img.grayed-out-img {
   left: 50%;
   top: 50%;
   transform: translateX(-50%) translateY(-50%);
-
-
 }
